Organize analysis viewers by domain
