+++ src/viewers/ssa/SSAViewer.tsx
@@ -1,9 +1,10 @@
-import { CodeViewer } from "./CodeViewer.tsx";
+import { CodeViewer } from "../CodeViewer.tsx";
 import React from "react";
 import { IPosition } from "monaco-editor";
-import { ViewerTitle } from "./ViewerTitle.tsx";
-import { VIEWER_TITLE_HEIGHT } from "../constants.ts";
-import { formatSSA, SSAFunction } from "../analysis.ts";
+import { ViewerTitle } from "../ViewerTitle.tsx";
+import { VIEWER_TITLE_HEIGHT } from "../../constants.ts";
+import { SSAFunction } from "../../analysis.ts";
+import { formatSSA } from "./format.ts";
 
 interface SSAViewerProps {
   src: string;
